requirements.txt
ZSI==2.0-rc3
argparse==1.2.1
beautifulsoup4==4.3.2
dateutils==0.6.6
distribute==0.6.27
html5lib==1.0b3
lxml==3.3.3
python-dateutil==2.2
pytz==2014.3
six==1.6.1
